Major code restructuring
The City class is now a primary class that keeps references to the
network layer, GIS database and event list inside it. Any object
that requires access to these features should store a pointer to
the City object in use.

This would also allow a single simulation to work with multiple
cities, by having separate network layers, geographic databases
and events for each city.

There is also a new routine to locate a vehicle by its GIS ID
(in addition to the identical one but for FCD IDs), and the
variable names throughout the simulation were made more distinct,
as the Swift compiler at this stage can easily get confused.