Performance improvements with the Haversine formula
This commit implements a Haversine formula for determining the
distance between two points. After profiling the simulator, we
found that >30% of the simulation time was spent in the routines
GIS.getDistance(fromPoint) and GIS.getFeatureGIDs(inCircleWithRadius).
By replacing the former with the Haversine simplification and the
latter with an in-simulation search, the overall simulation time
is pared down significantly.