A set of scripts to generate model-following floating car data The main script here, interact.py, connects to a SUMO instance and controls the insertion and removal of vehicles. It includes a parametric model for parking that follows real-life times of parking in the city of Chicago.
The script requires a clean SUMO net file, as it has no way to fail
gracefully when an invalid route is randomly selected. This occurs
primarily at fringe edges with a single direction, that cannot serve
as route destinations. One such map is included for the city of Porto.

A Dockerfile is included for a SUMO installation that listens to a
port and automatically restarts after Python/TraCI is finished.